api/requirements.txt
flask
flask-cors
huggingface-hub
mysql-connector-python
gunicorn
